+++ plotnine/scales/scale_stroke.py
@@ -1,7 +1,10 @@
+from warnings import warn
+
 import numpy as np
 from mizani.palettes import rescale_pal
 
 from ..doctools import document
+from ..exceptions import PlotnineWarning
 from ..utils import alias
 from .scale import scale_discrete, scale_continuous
 
@@ -26,7 +29,7 @@ class scale_stroke_continuous(scale_continuous):
 
 
 @document
-class scale_stroke_discrete(scale_discrete):
+class scale_stroke_ordinal(scale_discrete):
     """
     Discrete Stroke Scale
 
@@ -45,6 +48,25 @@ class scale_stroke_discrete(scale_discrete):
 
         self.palette = palette
         scale_discrete.__init__(self, **kwargs)
+
+
+@document
+class scale_stroke_discrete(scale_stroke_ordinal):
+    """
+    Discrete Stroke Scale
+
+    Parameters
+    ----------
+    {superclass_parameters}
+    """
+    _aesthetics = ['stroke']
+
+    def __init__(self, **kwargs):
+        warn(
+            "Using stroke for a ordinal variable is not advised.",
+            PlotnineWarning
+        )
+        super().__init__(self, **kwargs)
 
 
 alias('scale_stroke', scale_stroke_continuous)